Rename merchant route to merchant-dashboard and update payment demo flows
