+++ app/api/errors/run/route.ts
@@ -80,7 +80,7 @@ function scenarioCatalog(): Record<ScenarioId, ScenarioDefinition> {
       ].join('\n'),
       recoverySnippet: [
         'await money.faucet({ chain: "fast", network: "testnet" });',
-        'await money.send({ chain: "fast", network: "testnet", to: "<valid set1...>", amount: "1" });',
+        'await money.send({ chain: "fast", network: "testnet", to: "<valid fast1...>", amount: "1" });',
       ].join('\n'),
       runFail: async () => {
         const setup = await money.setup({ chain: 'fast', network: 'testnet' });
