+++ app/api/pay/route.ts
@@ -6,7 +6,7 @@ export const dynamic = 'force-dynamic';
 
 // Address validation patterns
 const PATTERNS: Record<string, RegExp> = {
-  fast: /^set1[a-z0-9]{38,}$/,
+  fast: /^(set|fast)1[a-z0-9]{38,}$/,
   evm: /^0x[0-9a-fA-F]{40}$/,
   solana: /^[1-9A-HJ-NP-Za-km-z]{32,44}$/,
 };
