+++ app/lib/demo/payment-links/native.ts
@@ -34,7 +34,7 @@ export const nativePaymentLinkProvider: PaymentLinkProviderAdapter = {
   id: 'native',
   mode: 'hosted-checkout',
   async createLink(input) {
-    const url = new URL('/merchant/checkout', input.baseUrl);
+    const url = new URL('/merchant-dashboard/checkout', input.baseUrl);
     url.searchParams.set('intentId', input.intentId);
     return { url: url.toString() };
   },
